Updates to decoding and frame unit tests.
After re-reading the spec, the only utf-8 content in the protocol is the actual content of the shortstr or longstr, update the decode, frame, header and codegen to support this.

Dont consume bytes on bit decoding until we've used all 8 bytes or we've moved from a bit field to a non-bit field.

Add frame tests auto-generated by monkey-patching the pika 0.9.6pr0 examples.

Still need to update encoding to support the bit syntax and to support the non-utf-8 content everywhere but in the actual string content for shortstr and longstr.